+++ frontend/src/main.jsx
@@ -2,6 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import './theme.css'
+import './print.css'
 import './App.css'
 import { ThemeProvider } from './theme/ThemeContext.jsx'